+++ d3d8/cpp/mge/inidata.h
@@ -59,6 +59,8 @@ const iniSetting iniSettings[] = {
         0),
     BIT(Configuration.MGEFlags, USE_HW_SHADER_BIT, "render.enable_shaders", 0),
     NUM(Configuration.EnableIndexedSkinning, t_bool, "render.indexed_skinning", 0),
+    NUM(Configuration.EnableCameraRelativeRendering, t_bool, "render.camera_relative", 0),
+    NUM(Configuration.CameraRelativeProbe, t_bool, "render.camera_relative_probe", 0),
     NUM(Configuration.HDRReactionSpeed, t_float, "render.hdr_reaction_time", 0),
     NUM(Configuration.PerPixelLightFlags, t_uint32, "distant_land.per_pixel_mode", 0),